Apply date validaror per field
Refactor dates validator so that there is one validator which is
applied per field and checks that the given date is after today.
Moreover, there is one more validator which is applied on an iterable
of pairs (start date and end date) and checks that end date is indeed
after start date.