Hide user time/tries user fields in django admin
The fields are to be used internally to track emails sent and avoid
DoS attacks, so they shouldn't appear in user creation or modification.

The admin loses the ability to reset a user and allow him to change their
password or verify but it was a hacky way too to achieve that. Let them
wait or find another way.